Add iOS 13 Swift template.  Lots of config and xcuitest cleanup.  Update to XCode 11.2.1.
